@@ -2,22 +2,39 @@ import click
 import os
 import subprocess
 
+
 FILE = os.path.abspath(__file__)
 PACKAGE_DIR = os.path.dirname(FILE)
 
 
 @click.command(context_settings=dict(ignore_unknown_options=True))
 @click.argument("method", type=click.Choice(["atac", "chip", "rna", "snp"]))
-@click.option("-g", "--genome", default="other", help="Genome to use",
-              type=click.Choice(choices=['dm6', 'hg19', 'hg38', 'hg38_dm6', 'hg38_mm39', 'hg38_spikein', 'mm10', 'mm39', 'other']),)
-
+@click.option(
+    "-g",
+    "--genome",
+    default="other",
+    help="Genome to use",
+    type=click.Choice(
+        choices=[
+            "dm6",
+            "hg19",
+            "hg38",
+            "hg38_dm6",
+            "hg38_mm39",
+            "hg38_spikein",
+            "mm10",
+            "mm39",
+            "other",
+        ]
+    ),
+)
 def cli_config(method, help=False, genome="other"):
     """
     Runs the config for the data processing pipeline.
     """
     import seqnado.config as config
-    config.create_config(method, genome)
 
+    config.create_config(method, genome)
 
 
 @click.command()
@@ -28,25 +45,31 @@ def cli_design(method, files, output="design.csv"):
     """
     Generates a SeqNado design file from a list of files.
     """
+    import pathlib
+    import sys
+    from seqnado.utils import Design, DesignIP, FastqFile, FastqFileIP
 
-    assert (
-        len(files) > 0
-    ), "No files provided. Please provide a list of files separated by spaces."
+    if not files:
+        files = list(pathlib.Path(".").glob("*.fastq.gz"))
+
+        if not files:
+            raise ValueError("No fastq files provided or found in current directory.")
 
     if not method == "chip":
-        from seqnado.utils import GenericFastqSamples
-
-        design = GenericFastqSamples.from_files(files).design
+        design = Design.from_fastq_files([FastqFile(path=fq) for fq in files])
     else:
-        from seqnado.utils import ChipseqFastqSamples
+        from seqnado.utils import DesignIP
 
-        design = ChipseqFastqSamples.from_files(files).design
+        design = DesignIP.from_fastq_files([FastqFileIP(path=fq) for fq in files])
 
-    design = design.drop(columns=["paired"], errors="ignore")
-    design.to_csv(output, index=False)
+    design.to_dataframe().to_csv(output)
 
 
 @click.command(context_settings=dict(ignore_unknown_options=True))
+@click.argument(
+    "method",
+    type=click.Choice(["atac", "chip", "rna", "snp", "consensus-peaks"]),
+)
 @click.argument(
     "method",
     type=click.Choice(["atac", "chip", "rna", "snp", "consensus-peaks"]),
@@ -67,11 +90,12 @@ def cli_design(method, files, output="design.csv"):
 def cli_pipeline(
     method, pipeline_options, help=False, cores=1, preset="local", version=False
 ):
-
     """Runs the data processing pipeline"""
 
     if version:
         from importlib.metadata import version
+
+        _version = version("seqnado")
 
         _version = version("seqnado")
         print(f"SeqNado version {_version}")
@@ -111,8 +135,11 @@ def cli_pipeline(
             ]
         )
 
+    cmd.extend(["--show-failed-logs"])
+
     with open(f"{PACKAGE_DIR}/data/logo.txt", "r") as f:
         logo = f.read()
 
     print(logo)
+
     completed = subprocess.run(cmd)
